Reorganize header file structure.
